chore: Update in-app links CUTR-at-USF->MobilityData
In several parts of the application URLs to the CUTR repo still existed:
1. For downloading unit test data
2. In webapp links (e.g., clicking on the error code in the error message to go to RULES.md)
3. In logs

I've updated these to point to the MobilityData repo instead.

Note that I've left other URLs in comments pointing at the CUTR repo, because those often reference specific issues and comments relevant to the history of certain changes.